backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-auth==2.27.0
google-cloud-dlp==3.15.0
google-cloud-aiplatform>=1.43.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
sse-starlette==1.8.2